+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## Release 1.2.0
+
+### New Features
+
+- api: add inverse parameter to coverage endpoint ([#762])
+- building-comparison: add caching of requests made to the database for EUBUCCO buidings ([#754])
+- building-comparison: add indicator to the list of core indicators and as such make it available on the ohsome Dashboard ([#765])
+
+[#754]: https://github.com/GIScience/ohsome-quality-api/pull/754
+[#762]: https://github.com/GIScience/ohsome-quality-api/issues/762
+[#765]: https://github.com/GIScience/ohsome-quality-api/issues/765
+
 ## Release 1.1.1
 
 ### Bug Fixes